+++ skword2vec/streams/__init__.py
@@ -1,0 +1,1 @@
+from skword2vec.streams._stream import Stream
